+++ lib/iex/api/crypto.rb
@@ -1,0 +1,13 @@
+module IEX
+  module Api
+    module Crypto
+      def self.get
+        connection.get.body
+      end
+
+      def self.connection
+        IEX::Api.default_connection 'market/crypto'
+      end
+    end
+  end
+end